move get_status from lockstatus.py to misc.py
Because it is generally useful and lockstatus.py only has that.
