+++ teuthology/provision.py
@@ -7,7 +7,7 @@ import yaml
 from .config import config
 from .contextutil import safe_while
 from .misc import decanonicalize_hostname, get_distro, get_distro_version
-from .lockstatus import get_status
+import misc
 
 
 log = logging.getLogger(__name__)
@@ -46,7 +46,7 @@ class Downburst(object):
         self.name = name
         self.os_type = os_type
         self.os_version = os_version
-        self.status = status or get_status(self.name)
+        self.status = status or misc.get_status(self.name)
         self.config_path = None
         self.host = decanonicalize_hostname(self.status['vm_host']['name'])
         self.executable = downburst_executable()
@@ -181,7 +181,7 @@ def create_if_vm(ctx, machine_name, _downburst=None):
     if _downburst:
         status_info = _downburst.status
     else:
-        status_info = get_status(machine_name)
+        status_info = misc.get_status(machine_name)
     if not status_info.get('is_vm', False):
         return False
     os_type = get_distro(ctx)
@@ -210,7 +210,7 @@ def destroy_if_vm(ctx, machine_name, user=None, description=None,
     if _downburst:
         status_info = _downburst.status
     else:
-        status_info = get_status(machine_name)
+        status_info = misc.get_status(machine_name)
     if not status_info or not status_info.get('is_vm', False):
         return True
     if user is not None and user != status_info['locked_by']:
